js/modules/selector: sepparate selector to selector and selectorUnit logic-wise
logically, it makes more sense that a Selector would be a class and each
SelectorUnit would be a class within the Selector class, because on both
we would like to be able to run different manipulations.
e.g. Selector -> "A#id1, B:class2"
e.g. SelectorUnit -> "A#id1" and "B:class2"